nlp/requirements.txt
numpy~=2.1.3
python-dotenv~=1.0.1
faiss-cpu~=1.9.0.post1
azure-core~=1.32.0
azure-identity
azure-servicebus
azure-appconfiguration
azure-keyvault-secrets
sentence-transformers~=3.3.1
fastapi~=0.115.6
pydantic~=2.10.3
openai~=1.57.1
transformers~=4.46.3